+++ dev/e2e_app/lib/keys.dart
@@ -36,6 +36,7 @@ class Keys {
   static const icon1 = Key('icon1');
   static const icon2 = Key('icon2');
   static const cameraFeaturesButton = Key('cameraFeaturesButton');
+  static const openExternalAppScreenButton = Key('openExternalAppScreenButton');
 
   // text fields screen
   static const textField1 = Key('textField1');
@@ -87,6 +88,10 @@ class Keys {
   static const pickMultiplePhotosButton = Key('pickMultiplePhotosButton');
   static const smallImagePreview = Key('smallImagePreview');
   static const selectedPhotosCount = Key('selectedPhotosCount');
+
+  // external app screen
+  static const openIosSettingsButton = Key('openIosSettingsButton');
+  static const externalAppStatusText = Key('externalAppStatusText');
 }
 
 class AtFinderItemKey extends ValueKey<int> {